feat: moves model classes to android module
