+++ android/src/main/java/com/rudderstack/android/models/RudderTraits.java
@@ -1,19 +1,4 @@
-/*
- * Creator: Debanjan Chatterjee on 15/12/21, 3:22 PM Last modified: 14/12/21, 8:22 PM
- * Copyright: All rights reserved Ⓒ 2021 http://rudderstack.com
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License. You may obtain a
- * copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-package com.rudderstack.models.android;
-
+package com.rudderstack.android.models;
 
 import com.google.gson.Gson;
 import com.google.gson.annotations.SerializedName;
